pyproject.toml
[project]
name = "tensor-ready-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-text-splitters>=0.3.9",
    "pydantic>=2.11.7",
    "pypdf>=5.9.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
